+++ examples/nextjs/pages/_app.tsx
@@ -1,0 +1,7 @@
+import "../styles/globals.css";
+import "../styles/base.css";
+import type { AppProps } from "next/app";
+
+export default function App({ Component, pageProps }: AppProps) {
+	return <Component {...pageProps} />;
+}