+++ test-apps/sveltekit/src/routes/+layout.server.ts
@@ -1,5 +1,4 @@
-import { auth } from '$lib/server/lucia';
 import { handleServerSession } from '@lucia-auth/sveltekit';
 import type { LayoutServerLoad } from './$types.js';
 
-export const load: LayoutServerLoad = handleServerSession(auth);
+export const load: LayoutServerLoad = handleServerSession();