+++ packages/express/CHANGELOG.md
@@ -1,1 +1,5 @@
 # CHANGELOG
+
+## 0.1.1
+
+- Add `locals.getSessionUser()` method