+++ packages/lucia-auth/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.1.2
+
+- [Breaking] `validateRequest()` requires `setCookie()` parameter
+- Add `getSessionUserFromRequest()`
+
 ## 0.1.1
 
 - Make input type of request for `parseRequest()` and `validateRequest()` as minimal as possible